requirements.txt
pandas
pyyaml
google-api-python-client
google-auth-oauthlib
python-dotenv
pytz
tqdm
plotly
xlsxwriter
pymongo
pydantic
